prevent error when trying to delete non-existing files

+++ classes/observer.php
@@ -123,8 +123,9 @@ class observer {
                     } else {
                         // File has been removed from assign.
                         // Remove from publication (file and db entry).
-                        $file = $fs->get_file_by_id($oldpubfile->fileid);
-                        $file->delete();
+                        if ($file = $fs->get_file_by_id($oldpubfile->fileid)) {
+                            $file->delete();
+                        }
 
                         $conditions['id'] = $oldpubfile->id;
